Fix incorrect checks on the existance of the Lock-Token header in responses
For NetFx the GetValues method was used which throws in the requested values are not found, which defaults the purpose of the fallback on the lock token in the response content.

The Uwp implementation always threw in the LockToken constructor since it was only invoked if the header was not present.